Allow tests to disable implicit permissive mode tests
